Added new storage selector
