+++ lib/screens/DownloadLocationsSettingsScreen.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 
 import '../components/DownloadLocationSettingsScreen/DownloadLocationList.dart';
-import '../components/DownloadLocationSettingsScreen/AddDownloadLocationDialog.dart';
-import '../components/DownloadLocationSettingsScreen/NewAddDownloadLocationDialog.dart';
 
 class DownloadsSettingsScreen extends StatelessWidget {
   const DownloadsSettingsScreen({Key key}) : super(key: key);
@@ -12,22 +10,11 @@ class DownloadsSettingsScreen extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(
         title: Text("Download Locations"),
-        actions: [
-          IconButton(
-            icon: Icon(Icons.sd_storage),
-            onPressed: () => showDialog(
-              context: context,
-              builder: (context) => NewAddDownloadLocationDialog(),
-            ),
-          )
-        ],
       ),
       floatingActionButton: FloatingActionButton(
         child: Icon(Icons.add),
-        onPressed: () => showDialog(
-          context: context,
-          builder: (context) => AddDownloadLocationDialog(),
-        ),
+        onPressed: () =>
+            Navigator.of(context).pushNamed("/settings/downloadlocations/add"),
       ),
       body: DownloadLocationList(),
     );
