+++ pubspec.yaml
@@ -44,6 +44,7 @@ dependencies:
   clipboard: ^0.1.3
   file_picker: ^2.1.7
   permission_handler: ^5.1.0+2
+  provider: ^5.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
